Mapping modules
The `map` config option provided by RequireJS allows us to have two
different names for a module. You might get this problem when your
module depends on a thirdparty module. This might be better expliained
with and example.

Lets say you have a module that depends on 'mustache'. Now your module
needs to be consumed by an application that places its thirdparty
dependencies in a _lib_ directory. When this applications modules depend
on mustache, they depend on 'lib/mustache'.

When RequireJS is trying to resolve your 'mustache' dependency it's not
looking in the _lib_ directory so it fails. Using `map` can help here to
provide multiple module names for the same file.

This is not really the point of the `map` config option. Its more so you
can play with multiple versions of the same library. e.g. _Module A_
depends on version 1 of _Module C_, but _Module B_ depends on _C 2.0_.